Moved settings section view models into dedicated namespace
