+++ src/Smallify.Module.Settings/ViewModels/Sections/AboutViewModel.cs
@@ -4,7 +4,7 @@ using Smallify.Module.Core;
 using System.Diagnostics;
 using System.Windows.Input;
 
-namespace Smallify.Module.Settings.ViewModels
+namespace Smallify.Module.Settings.ViewModels.Sections
 {
 	public class AboutViewModel : BindableBase, IAboutViewModel
 	{
