+++ src/Smallify.Module.Settings/ViewModels/Sections/GeneralViewModel.cs
@@ -1,7 +1,7 @@
 ﻿using Prism.Mvvm;
 using Smallify.Module.Core;
 
-namespace Smallify.Module.Settings.ViewModels
+namespace Smallify.Module.Settings.ViewModels.Sections
 {
 	public class GeneralViewModel : BindableBase, IGeneralViewModel
 	{
