+++ src/Smallify.Module.Settings/ViewModels/Sections/IGeneralViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Smallify.Module.Settings.ViewModels
+﻿namespace Smallify.Module.Settings.ViewModels.Sections
 {
 	public interface IGeneralViewModel
 	{
